Reduced Simulation Frame Size
* Merge

* Rebase

* Tidy ups

* hate this file

* Optimal

* Merge branch 'chore/reduce-cpu-pressure' into chore/reduced-frame-size

* Wip

* Remove stagger

* revert dlls

* Merge branch 'chore/reduce-cpu-pressure' into chore/reduced-frame-size

* Fix

* Tidy ups

* Tidy ups

* Merge branch 'develop' into chore/reduced-frame-size

* Fix test

* Fix tests

* Fix tests

* Fix test


Approved-by: Matt Ryan

+++ BlastRoyale/Assets/Src/FirstLight/Game/Commands/UpgradeItemCommand.cs
@@ -41,7 +41,7 @@ namespace FirstLight.Game.Commands
 				GameId = info.Equipment.GameId,
 				Name = info.Equipment.GameId.ToString(),
 				Durability = info.CurrentDurability,
-				Level = item.Level + 1,
+				Level = (uint)(item.Level + 1),
 				Price = cost
 			});
 			return UniTask.CompletedTask;
